DI + Create sessions a bit earlier in the sequence
PlayerJoinEvent - "Called when the player spawns in the world after logging in, when they first see the terrain."

PlayerLoginEvent - Called after the player has successfully authenticated, before it spawns. The player is on the loading screen when this is called.

This would (in most optimal situations) allow the member data to be available upon PlayerJoinEvent